fix(linux): fix prepare
